+++ Gemfile
@@ -1,7 +1,10 @@
 # frozen_string_literal: true
 source "https://rubygems.org"
 
+gem 'activerecord'
 gem "sinatra-activerecord"
 gem "sqlite3"
 gem "pry"
 gem "require_all"
+gem 'rest-client'
+
